Support multiple model versions
New option "model_versions: VERSION" allows applications to use a
specific model version. VERSION can be either of 0.149 or 0.151
for now.

Future implementations may need to do a smart way if VERSION doesn't
exist. For example, if 0.150 is set, use 0.149 (older latest version).
